+++ pychunkedgraph/exporting/export.py
@@ -4,7 +4,7 @@ import itertools
 import pickle as pkl
 
 from pychunkedgraph.backend import chunkedgraph
-from pychunkedgraph.multiprocessing import multiprocessing_utils as mu
+from pychunkedgraph.paralellizing import multiprocessing_utils as mu
 
 
 def get_sv_to_root_id_mapping_chunk(cg, chunk_coords, vol=None):
@@ -158,7 +158,7 @@ def write_flat_segmentation(cg, dataset_name, bounding_box=None, block_factor=2,
         multi_args.append([cg_info, start_block, end_block,
                            from_url, to_url, mip])
 
-    # Run multiprocessing
+    # Run paralellizing
     if n_threads == 1:
         mu.multiprocess_func(_write_flat_segmentation_thread, multi_args,
                              n_threads=n_threads, verbose=True,
